move safeParseArray to own file, update imports
